+++ sc-memory/cpp/python/sc_python_module.cpp
@@ -232,6 +232,12 @@ public:
     return value;
   }
 
+  bp::object AsBinary() const
+  {
+    PyObject * buff = PyMemoryView_FromMemory((char*)m_buffer->Data(), m_buffer->Size(), PyBUF_READ);
+    return boost::python::object(boost::python::handle<>(buff));
+  }
+
 private:
   MemoryBufferSafePtr m_buffer;
   uint8_t m_type;
@@ -748,6 +754,7 @@ BOOST_PYTHON_MODULE(sc)
     .def("AsString", &impl::PyLinkContent::AsString)
     .def("AsInt", &impl::PyLinkContent::AsInt)
     .def("AsFloat", &impl::PyLinkContent::AsDouble)
+    .def("AsBinary", &impl::PyLinkContent::AsBinary)
     .def("GetType", &impl::PyLinkContent::GetType)
     .def_readonly("String", &impl::PyLinkContent::Type::String)
     .def_readonly("Int", &impl::PyLinkContent::Type::Int)