+++ docs/changelog.md
@@ -2,7 +2,7 @@
 
 * Add `ScSet` and `ScRelationSet` in python common library
 * Add `ScPythonService` into C++ (allow to implement threaded python services)
-* Add `GetType` function into Python for a link content
+* Add `GetType` and `AsBinary` functions into Python for a link content
 * Add colors to python output
 * Fix bug with python modules search path initialization
 * Fix bug with python threading